Projectiles + Player Punches
Projectiles in.   See User_Controllable for the particulars of how to create and send on its way.

Player punches are in as well.  They should be fast enough now that no matter how fast the player is, they'll never run into their own punch.  No room for ricochet, but that's just an out of scope matter.